requirements.txt
PyQt5
openai
mss
Pillow